requirements.txt
Flask==2.2.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.4.4
PyJWT~=2.8.0
Werkzeug~=2.3.6
google-cloud-firestore==2.9.0
python-dotenv==1.0.0
redis==4.2.2
flask-caching==2.3.0